[Rules] Adds a RulesDefinition.
And a very basic 'rules' file. This rules file will serve two
purposes : it is used to define the RulesRepository (static
list of rules made available by this plugin) and it will be used
to generate a .ndproj from the .ndproj template.
Also, changes the sonar API from version 3.7 to version 4.5

Change-Id: I4ec8950496f15d749cb081e06e7f4e69a25cb673